+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- `src` for runner
+- `_tests_` for runner
+
 ## 0.0.1 - 2024-11-20
 ### Added
 - Project init and first publish